Move pkg/streamformatter to api/pkg/streamformatter
