+++ api/go.mod
@@ -5,11 +5,10 @@ go 1.23.0
 require (
 	github.com/docker/go-connections v0.5.0
 	github.com/docker/go-units v0.5.0
+	github.com/google/go-cmp v0.5.9
 	github.com/moby/docker-image-spec v1.3.1
 	github.com/opencontainers/go-digest v1.0.0
 	github.com/opencontainers/image-spec v1.1.1
 	golang.org/x/time v0.11.0
 	gotest.tools/v3 v3.5.2
 )
-
-require github.com/google/go-cmp v0.5.9 // indirect
