+++ api/pkg/streamformatter/streamformatter_test.go
