+++ api/pkg/streamformatter/streamwriter.go
