+++ api/pkg/streamformatter/streamwriter_test.go
