+++ daemon/server/router/plugin/plugin_routes.go
@@ -10,7 +10,7 @@ import (
 	"github.com/docker/docker/daemon/server/backend"
 	"github.com/docker/docker/daemon/server/httputils"
 	"github.com/docker/docker/pkg/ioutils"
-	"github.com/docker/docker/pkg/streamformatter"
+	"github.com/moby/moby/api/pkg/streamformatter"
 	"github.com/moby/moby/api/types"
 	"github.com/moby/moby/api/types/filters"
 	"github.com/moby/moby/api/types/registry"
